rename sentence -> related
